+++ test/unit/core/test_instructions.py
@@ -39,9 +39,18 @@ def _build_and_get_matrix(gate_func: Callable | tuple[Callable, Callable], *args
         (instructions.z, SINGLE_QUBIT_GATE_MATRICES["Z"]),
         (instructions.t, SINGLE_QUBIT_GATE_MATRICES["T"]),
         (instructions.t_dag, SINGLE_QUBIT_GATE_MATRICES["T_DAG"]),
+        (instructions.c_nxyz, SINGLE_QUBIT_GATE_MATRICES["C_NXYZ"]),
+        (instructions.c_nzyx, SINGLE_QUBIT_GATE_MATRICES["C_NZYX"]),
+        (instructions.c_xnyz, SINGLE_QUBIT_GATE_MATRICES["C_XNYZ"]),
+        (instructions.c_xynz, SINGLE_QUBIT_GATE_MATRICES["C_XYNZ"]),
         (instructions.c_xyz, SINGLE_QUBIT_GATE_MATRICES["C_XYZ"]),
+        (instructions.c_znyx, SINGLE_QUBIT_GATE_MATRICES["C_ZNYX"]),
+        (instructions.c_zynx, SINGLE_QUBIT_GATE_MATRICES["C_ZYNX"]),
         (instructions.c_zyx, SINGLE_QUBIT_GATE_MATRICES["C_ZYX"]),
         (instructions.h, SINGLE_QUBIT_GATE_MATRICES["H"]),
+        (instructions.h_nxy, SINGLE_QUBIT_GATE_MATRICES["H_NXY"]),
+        (instructions.h_nxz, SINGLE_QUBIT_GATE_MATRICES["H_NXZ"]),
+        (instructions.h_nyz, SINGLE_QUBIT_GATE_MATRICES["H_NYZ"]),
         (instructions.h_xy, SINGLE_QUBIT_GATE_MATRICES["H_XY"]),
         (instructions.h_yz, SINGLE_QUBIT_GATE_MATRICES["H_YZ"]),
         (instructions.s, SINGLE_QUBIT_GATE_MATRICES["S"]),
@@ -88,8 +97,10 @@ def test_u3(frac_theta: Fraction, frac_phi: Fraction, frac_lambda: Fraction):
     "gate_func, matrix",
     [
         (instructions.cnot, TWO_QUBIT_GATE_MATRICES["CNOT"]),
+        (instructions.cxswap, TWO_QUBIT_GATE_MATRICES["CXSWAP"]),
         (instructions.cy, TWO_QUBIT_GATE_MATRICES["CY"]),
         (instructions.cz, TWO_QUBIT_GATE_MATRICES["CZ"]),
+        (instructions.czswap, TWO_QUBIT_GATE_MATRICES["CZSWAP"]),
         (instructions.iswap, TWO_QUBIT_GATE_MATRICES["ISWAP"]),
         (instructions.iswap_dag, TWO_QUBIT_GATE_MATRICES["ISWAP_DAG"]),
         (instructions.sqrt_xx, TWO_QUBIT_GATE_MATRICES["SQRT_XX"]),
@@ -99,6 +110,8 @@ def test_u3(frac_theta: Fraction, frac_phi: Fraction, frac_lambda: Fraction):
         (instructions.sqrt_zz, TWO_QUBIT_GATE_MATRICES["SQRT_ZZ"]),
         (instructions.sqrt_zz_dag, TWO_QUBIT_GATE_MATRICES["SQRT_ZZ_DAG"]),
         (instructions.swap, TWO_QUBIT_GATE_MATRICES["SWAP"]),
+        (instructions.swapcx, TWO_QUBIT_GATE_MATRICES["SWAPCX"]),
+        (instructions.swapcz, TWO_QUBIT_GATE_MATRICES["SWAPCZ"]),
         (instructions.xcx, TWO_QUBIT_GATE_MATRICES["XCX"]),
         (instructions.xcy, TWO_QUBIT_GATE_MATRICES["XCY"]),
         (instructions.xcz, TWO_QUBIT_GATE_MATRICES["XCZ"]),
@@ -171,7 +184,7 @@ def test_m():
     assert np.allclose(result, id / np.sqrt(2))
 
 
-def tes_mx():
+def test_mx():
     id = np.eye(2)
     result = _build_and_get_matrix(instructions.mx, 0)
     assert np.allclose(result, id / np.sqrt(2))